app/requirements.txt
Flask==1.0.2
Keras==2.2.4
numpy==1.16.2
pandas==0.24.1
scikit-learn==0.20.3
tensorflow==1.13.1
mysql-connector
tweepy==3.7.0
textblob==0.15.3
matplotlib==3.0.3
nltk==3.4.1

